Implement Scala Native parser using jawn
