+++ play-json/js-native/src/test/scala-3/play/api/libs/json/EnumSpec.scala
