+++ play-json/native/src/test/scala/java/security/SecureRandom.scala
@@ -1,0 +1,4 @@
+package java.security
+
+// insecure implementation only for tests
+class SecureRandom extends java.util.Random
